src/functions/fix_image_urls/requirements.txt
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.23.1
pandas==2.2.3